pyproject.toml
[project]
name = "openapi-to-cli"
authors = [{ name = "Joshua Oliphant", email = "joshua.oliphant@hey.com" }]
version = "0.1.0"
description = "A CLI generator for OpenAPI specs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "jinja2>=3.1.4",
    "openapi-python-client>=0.21.5",
    "pytest>=8.3.3",
    "pyyaml>=6.0.2",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
